[airflow] Support SNS alerts
Summary:
Airflow only natively supports smtp email, but we can connect it to SNS.
This just sets it up with a topic that hits me, and I can later set up
the lambda which writes to slack.

Test Plan:
Purposefully made a job fail (SQL syntax error) and it sends an SNS
notification to my email.

Reviewers: hcli, ilan, keoki

Reviewed By: keoki

Differential Revision: https://openeducationlabs.phacility.com/D8694