Availability monitoring
Summary:
I want to track our availability: how many "nines" of availability do
we have. To do this, I will query the warehouse with a reference query
every minute and record the results (starttime, endtime) to a table.
With that table, I will track how often the query completes in <10s and
count that as our availability number as a Mode report.

Test Plan:
Run the redshift_availability_minutely job and check that it shows up in
the query log.
Run the redshift_availability_rollup job and check that the
quality_of_service.availability_job picks up the data.

Reviewers: hcli, ilan, keoki

Reviewed By: ilan

Differential Revision: https://openeducationlabs.phacility.com/D8778