Query killer with SNS notification and log
Summary:
Kills queries which exceed a certain running time (excluding table loads, since these can take a while legitimately). BTW @shenny did this today and the warehouse was totally slow, so there's a need here.

On kill, it messages the slack channel. It also logs a snapshot of all running queries, slow or not, for auditing and investigating which queries consume the most resources.

Coincidentally, after writing this I discovered that WLM supports a max execution time which is equivalent to this. However, I liked that mine messages the slack channel for shaming and gives us a log which we can use to mine long-running queries in general.

Test Plan: Ran both jobs with dry_run=False

Reviewers: hcli, ilan, kristen

Reviewed By: ilan

Differential Revision: https://openeducationlabs.phacility.com/D8978