Fix LISTAGG in query_killer
Summary:
Only select <65535 characters. If we really want to see the whole query, we could go one step further and do the first 65400 characters in query1, the next 65400 in query2, etc., then join them. But not going that far yet.

Also add the user.

Test Plan: Manually test the query on query id 671526 which has a ridiculous size.

Reviewers: ilan, hcli

Reviewed By: ilan

Differential Revision: https://openeducationlabs.phacility.com/D9002